Scratch pad for Math::Model
The vision is that people can write down a simple model that corresponds to a
set of ordinarly, linear differential equations, and this module brushes it up
for you and passes it to Math::RungeKutta for integration, and produces some
nice looking charts.

It will take quite a while to get there (if at all), but so far the hacking is
fun.

See also: http://www.perlmonks.org/?node_id=645771